refactor: reorganize files for simpler lib
